pk's are now big integers

+++ models/UserListEntry.py
@@ -3,6 +3,6 @@ from peewee import *
 from base.BaseModel import BaseModel
 
 class UserListEntry( BaseModel ):
-    owner_pk = IntegerField( )
-    entry_pk = IntegerField( )
+    owner_pk = BigIntegerField( )
+    entry_pk = BigIntegerField( )
     ig_mode = TextField( )
